Basic read/write story + some assets + basics of ui
Added some assets as well as some ui elements that we can use to instansiate butons and such. Most importantly added some functions in StoryMaker where we can write json sorty, will eventually be used for users to write their own story. And a read story, that deserialised a json with a story in in, its barebones and doesnt iterate through the lists of thumbnails and choices yet, but thats simple enough. For now I want to be able to display a story into the UI, then work on being able to iterate throught the story choises and then work on grabbing json files with stories in them to read and display. On that note I would like for the write story to be output a json file, that way we can just grab it and read it ourselves. Then finally I can add a story making interface. Would be cool in the future to add file placement and possibly a sharing system - By cloud, bluetooth, QR code, idk look into it.