Added a simple story list + Scene builder should work a bit
Well the story lister is suddenly bugging for some reason, and the scene builder needs to build the choices and the choices need to send to the scene builder what scene the choice is linked to. Look at the story lister for how to do the choices, my brain is starting to melt rn so Hopefully I find some time this weekend or im cooked. Also look into adding the items, lets try and get at least 10/20...